Former-commit-id: 32aed401fe2d26f427bea4aebcd33e4f2cd3fd96

+++ Classik_Backend/src/main/kotlin/com/ssafy/Classik_Backend/repository/TrackRepository.kt
@@ -1,11 +1,11 @@
 package com.ssafy.Classik_Backend.repository
 
-import com.ssafy.Classik_Backend.domain.track.entity.Track
+import com.ssafy.Classik_Backend.entity.Track
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TrackRepository : JpaRepository<Track, Long> {
+interface TrackRepository : JpaRepository<Track, Int> {
 
     fun findAllByTitle(title: String): List<Track>
